Splitting the common helper + small refactor (#150)
Decent size refactor.
