+++ lib/db.js
@@ -11,6 +11,7 @@ function initDb(dbUrl, callback){ // eslint-disable-line
     MongoClient.connect(dbUrl, { useNewUrlParser: true, useUnifiedTopology: true }, connected);
     function connected(err, client){
         if(err){
+            console.log('Failed connecting to the DB', err);
             return callback(err);
         }
