+++ lib/indexing.js
@@ -82,7 +82,7 @@ const indexOrders = (app, cb) => {
     return new Promise((resolve, reject) => {
         app.db.orders.find({}).toArray((err, ordersList) => {
             if(err){
-                console.error(colors.red('Error setting up products index: ' + err));
+                console.error(colors.red(`Error setting up products index: ${err}`));
                 reject(err);
             }
