+++ lib/payments/adyen.js
@@ -1,12 +1,14 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
 const numeral = require('numeral');
 const { Client, CheckoutAPI } = require('@adyen/api-library');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const router = express.Router();
 
 router.post('/setup', async (req, res, next) => {
-    const adyenConfig = common.getPaymentConfig();
+    const adyenConfig = getPaymentConfig();
 
     const client = new Client({
         apiKey: adyenConfig.apiKey,
@@ -26,7 +28,7 @@ router.post('/setup', async (req, res, next) => {
         });
     }catch(ex){
         console.log('Exception getting supported payment methods', ex.message);
-        res.status(400).json({ message: 'Failed to retrieve payment methods.' + ex.message });
+        res.status(400).json({ message: `Failed to retrieve payment methods.${ex.message}` });
     }
     res.status(200).json({
         paymentsResponse,
@@ -38,7 +40,7 @@ router.post('/setup', async (req, res, next) => {
 router.post('/checkout_action', async (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const adyenConfig = common.getPaymentConfig();
+    const adyenConfig = getPaymentConfig();
 
     const client = new Client({
         apiKey: adyenConfig.apiKey,
@@ -80,7 +82,7 @@ router.post('/checkout_action', async (req, res, next) => {
         orderShipping: req.session.totalCartShipping,
         orderItemCount: req.session.totalCartItems,
         orderProductCount: req.session.totalCartProducts,
-        orderCustomer: common.getId(req.session.customerId),
+        orderCustomer: getId(req.session.customerId),
         orderEmail: req.session.customerEmail,
         orderCompany: req.session.customerCompany,
         orderFirstname: req.session.customerFirstname,
@@ -114,7 +116,7 @@ router.post('/checkout_action', async (req, res, next) => {
             req.session.message = 'Your payment was successfully completed';
             req.session.paymentEmailAddr = orderDoc.orderEmail;
             req.session.paymentApproved = true;
-            req.session.paymentDetails = '<p><strong>Order ID: </strong>' + newId + '</p><p><strong>Transaction ID: </strong>' + response.pspReference + '</p>';
+            req.session.paymentDetails = `<p><strong>Order ID: </strong>${newId}</p><p><strong>Transaction ID: </strong>${response.pspReference}</p>`;
 
             // set payment results for email
             const paymentResults = {
@@ -127,12 +129,12 @@ router.post('/checkout_action', async (req, res, next) => {
 
             // clear the cart
             if(req.session.cart){
-                common.emptyCart(req, res, 'function');
+                emptyCart(req, res, 'function');
             }
 
             // send the email with the response
             // TODO: Should fix this to properly handle result
-            common.sendEmail(req.session.paymentEmailAddr, 'Your payment with ' + config.cartTitle, common.getEmailTemplate(paymentResults));
+            sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
         }
         res.status(200).json({ paymentId: newId });
     });
