+++ lib/payments/authorizenet.js
@@ -1,15 +1,17 @@
 const express = require('express');
 const axios = require('axios');
 const stripBom = require('strip-bom');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const router = express.Router();
 
 // The homepage of the site
 router.post('/checkout_action', (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const authorizenetConfig = common.getPaymentConfig();
+    const authorizenetConfig = getPaymentConfig();
 
     let authorizeUrl = 'https://api.authorize.net/xml/v1/request.api';
     if(authorizenetConfig.mode === 'test'){
@@ -63,7 +65,7 @@ router.post('/checkout_action', (req, res, next) => {
             orderShipping: req.session.totalCartShipping,
             orderItemCount: req.session.totalCartItems,
             orderProductCount: req.session.totalCartProducts,
-            orderCustomer: common.getId(req.session.customerId),
+            orderCustomer: getId(req.session.customerId),
             orderEmail: req.session.customerEmail,
             orderCompany: req.session.customerCompany,
             orderFirstname: req.session.customerFirstname,
@@ -112,12 +114,12 @@ router.post('/checkout_action', (req, res, next) => {
 
                     // clear the cart
                     if(req.session.cart){
-                        common.emptyCart(req, res, 'function');
+                        emptyCart(req, res, 'function');
                     }
 
                     // send the email with the response
                     // TODO: Should fix this to properly handle result
-                    common.sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, common.getEmailTemplate(paymentResults));
+                    sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
                     // redirect to outcome
                     res.status(200).json({ orderId: newId });
