+++ lib/payments/blockonomics.js
@@ -1,6 +1,7 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
 const router = express.Router();
 const axios = require('axios').default;
 
@@ -35,12 +36,12 @@ router.get('/checkout_return', async (req, res, next) => {
                 messageType: 'success',
                 paymentEmailAddr: order.orderEmail,
                 paymentApproved: true,
-                paymentDetails: '<p><strong>Order ID: </strong>' + order._id + '</p><p><strong>Transaction ID: </strong>' + order.orderPaymentId + '</p>'
+                paymentDetails: `<p><strong>Order ID: </strong>${order._id}</p><p><strong>Transaction ID: </strong>${order.orderPaymentId}</p>`
             };
 
             // send the email with the response
             // TODO: Should fix this to properly handle result
-            common.sendEmail(req.session.paymentEmailAddr, 'Your payment with ' + config.cartTitle, common.getEmailTemplate(paymentResults));
+            sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
             res.status(200).json({ err: '' });
         }catch(ex){
             console.info('Error updating status success blockonomics', ex);
@@ -69,7 +70,7 @@ router.get('/checkout_return', async (req, res, next) => {
 });
 
 router.post('/checkout_action', (req, res, next) => {
-  const blockonomicsConfig = common.getPaymentConfig();
+  const blockonomicsConfig = getPaymentConfig();
   const config = req.app.config;
   const db = req.app.db;
   const blockonomicsParams = {};
@@ -80,7 +81,7 @@ router.post('/checkout_action', (req, res, next) => {
     blockonomicsParams.expectedBtc = Math.round(req.session.totalCartAmount / response.data.price * Math.pow(10, 8)) / Math.pow(10, 8);
     // get new address
     axios
-      .post(blockonomicsConfig.hostUrl + blockonomicsConfig.newAddressApi, {}, { headers: { 'Content-Type': 'application/json', 'User-Agent': 'blockonomics', Accept: 'application/json', Authorization: 'Bearer ' + blockonomicsConfig.apiKey } })
+      .post(blockonomicsConfig.hostUrl + blockonomicsConfig.newAddressApi, {}, { headers: { 'Content-Type': 'application/json', 'User-Agent': 'blockonomics', Accept: 'application/json', Authorization: `Bearer ${blockonomicsConfig.apiKey}` } })
       .then((response) => {
           blockonomicsParams.address = response.data.address;
           blockonomicsParams.timestamp = Math.floor(new Date() / 1000);
@@ -94,7 +95,7 @@ router.post('/checkout_action', (req, res, next) => {
                 orderShipping: req.session.totalCartShipping,
                 orderItemCount: req.session.totalCartItems,
                 orderProductCount: req.session.totalCartProducts,
-                orderCustomer: common.getId(req.session.customerId),
+                orderCustomer: getId(req.session.customerId),
                 orderEmail: req.session.customerEmail,
                 orderCompany: req.session.customerCompany,
                 orderFirstname: req.session.customerFirstname,
