+++ lib/payments/instore.js
@@ -1,23 +1,25 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const router = express.Router();
 
 // The homepage of the site
 router.post('/checkout_action', async (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const instoreConfig = common.getPaymentConfig();
+    const instoreConfig = getPaymentConfig();
 
     const orderDoc = {
-        orderPaymentId: common.getId(),
+        orderPaymentId: getId(),
         orderPaymentGateway: 'Instore',
         orderPaymentMessage: 'Your payment was successfully completed',
         orderTotal: req.session.totalCartAmount,
         orderShipping: 0,
         orderItemCount: req.session.totalCartItems,
         orderProductCount: req.session.totalCartProducts,
-        orderCustomer: common.getId(req.session.customerId),
+        orderCustomer: getId(req.session.customerId),
         orderEmail: req.session.customerEmail,
         orderCompany: req.session.customerCompany,
         orderFirstname: req.session.customerFirstname,
@@ -64,15 +66,15 @@ router.post('/checkout_action', async (req, res, next) => {
 
             // clear the cart
             if(req.session.cart){
-                common.emptyCart(req, res, 'function');
+                emptyCart(req, res, 'function');
             }
 
             // send the email with the response
             // TODO: Should fix this to properly handle result
-            common.sendEmail(req.session.paymentEmailAddr, `Your order with ${config.cartTitle}`, common.getEmailTemplate(paymentResults));
+            sendEmail(req.session.paymentEmailAddr, `Your order with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
             // redirect to outcome
-            res.redirect('/payment/' + newId);
+            res.redirect(`/payment/${newId}`);
         });
     }catch(ex){
         console.log('Error sending payment to API', ex);
