+++ lib/payments/paypal.js
@@ -1,6 +1,8 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const paypal = require('paypal-rest-sdk');
 const router = express.Router();
 
@@ -37,7 +39,7 @@ router.get('/checkout_return', (req, res, next) => {
             req.session.paymentApproved = paymentApproved;
             req.session.paymentDetails = paymentDetails;
 
-            res.redirect('/payment/' + req.session.orderId);
+            res.redirect(`/payment/${req.session.orderId}`);
             return;
         }
 
@@ -49,27 +51,27 @@ router.get('/checkout_return', (req, res, next) => {
             paymentApproved = true;
             paymentStatus = 'Paid';
             paymentMessage = 'Your payment was successfully completed';
-            paymentDetails = '<p><strong>Order ID: </strong>' + paymentOrderId + '</p><p><strong>Transaction ID: </strong>' + payment.id + '</p>';
+            paymentDetails = `<p><strong>Order ID: </strong>${paymentOrderId}</p><p><strong>Transaction ID: </strong>${payment.id}</p>`;
 
             // clear the cart
             if(req.session.cart){
-                common.emptyCart(req, res, 'function');
+                emptyCart(req, res, 'function');
             }
         }
 
         // failed
         if(payment.failureReason){
             paymentApproved = false;
-            paymentMessage = 'Your payment failed - ' + payment.failureReason;
+            paymentMessage = `Your payment failed - ${payment.failureReason}`;
             paymentStatus = 'Declined';
         }
 
         // update the order status
-        db.orders.updateOne({ _id: common.getId(paymentOrderId) }, { $set: { orderStatus: paymentStatus } }, { multi: false }, (err, numReplaced) => {
+        db.orders.updateOne({ _id: getId(paymentOrderId) }, { $set: { orderStatus: paymentStatus } }, { multi: false }, (err, numReplaced) => {
             if(err){
                 console.info(err.stack);
             }
-            db.orders.findOne({ _id: common.getId(paymentOrderId) }, (err, order) => {
+            db.orders.findOne({ _id: getId(paymentOrderId) }, (err, order) => {
                 if(err){
                     console.info(err.stack);
                 }
@@ -94,9 +96,9 @@ router.get('/checkout_return', (req, res, next) => {
 
                     // send the email with the response
                     // TODO: Should fix this to properly handle result
-                    common.sendEmail(req.session.paymentEmailAddr, 'Your payment with ' + config.cartTitle, common.getEmailTemplate(paymentResults));
+                    sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
-                    res.redirect('/payment/' + order._id);
+                    res.redirect(`/payment/${order._id}`);
                 });
             });
         });
@@ -107,7 +109,7 @@ router.get('/checkout_return', (req, res, next) => {
 router.post('/checkout_action', (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const paypalConfig = common.getPaymentConfig();
+    const paypalConfig = getPaymentConfig();
 
     // setup the payment object
     const payment = {
@@ -116,8 +118,8 @@ router.post('/checkout_action', (req, res, next) => {
             payment_method: 'paypal'
         },
         redirect_urls: {
-            return_url: config.baseUrl + '/paypal/checkout_return',
-            cancel_url: config.baseUrl + '/paypal/checkout_cancel'
+            return_url: `${config.baseUrl}/paypal/checkout_return`,
+            cancel_url: `${config.baseUrl}/paypal/checkout_cancel`
         },
         transactions: [{
             amount: {
@@ -166,7 +168,7 @@ router.post('/checkout_action', (req, res, next) => {
                 orderShipping: req.session.totalCartShipping,
                 orderItemCount: req.session.totalCartItems,
                 orderProductCount: req.session.totalCartProducts,
-                orderCustomer: common.getId(req.session.customerId),
+                orderCustomer: getId(req.session.customerId),
                 orderEmail: req.session.customerEmail,
                 orderCompany: req.session.customerCompany,
                 orderFirstname: req.session.customerFirstname,
