+++ lib/payments/payway.js
@@ -1,6 +1,8 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const numeral = require('numeral');
 const got = require('got');
 const randtoken = require('rand-token');
@@ -10,7 +12,7 @@ const router = express.Router();
 router.post('/checkout_action', async (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const paymentConfig = common.getPaymentConfig();
+    const paymentConfig = getPaymentConfig();
 
     // Create the payload
     const payload = {
@@ -61,7 +63,7 @@ router.post('/checkout_action', async (req, res, next) => {
             orderShipping: req.session.totalCartShipping,
             orderItemCount: req.session.totalCartItems,
             orderProductCount: req.session.totalCartProducts,
-            orderCustomer: common.getId(req.session.customerId),
+            orderCustomer: getId(req.session.customerId),
             orderEmail: req.session.customerEmail,
             orderCompany: req.session.customerCompany,
             orderFirstname: req.session.customerFirstname,
@@ -110,15 +112,15 @@ router.post('/checkout_action', async (req, res, next) => {
 
                     // clear the cart
                     if(req.session.cart){
-                        common.emptyCart(req, res, 'function');
+                        emptyCart(req, res, 'function');
                     }
 
                     // send the email with the response
                     // TODO: Should fix this to properly handle result
-                    common.sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, common.getEmailTemplate(paymentResults));
+                    sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
                     // redirect to outcome
-                    res.redirect('/payment/' + newId);
+                    res.redirect(`/payment/${newId}`);
                 }else{
                     // redirect to failure
                     req.session.messageType = 'danger';
@@ -126,7 +128,7 @@ router.post('/checkout_action', async (req, res, next) => {
                     req.session.paymentApproved = false;
                     req.session.paymentDetails = `<p><strong>Order ID: </strong>${newId}
                     </p><p><strong>Transaction ID: </strong> ${txn.transHash}</p>`;
-                    res.redirect('/payment/' + newId);
+                    res.redirect(`/payment/${newId}`);
                 }
             });
         }catch(ex){
