+++ lib/payments/stripe.js
@@ -1,15 +1,17 @@
 const express = require('express');
-const common = require('../common');
 const { indexOrders } = require('../indexing');
+const { getId, sendEmail, getEmailTemplate } = require('../common');
+const { getPaymentConfig } = require('../config');
+const { emptyCart } = require('../cart');
 const numeral = require('numeral');
-const stripe = require('stripe')(common.getPaymentConfig().secretKey);
+const stripe = require('stripe')(getPaymentConfig().secretKey);
 const router = express.Router();
 
 // The homepage of the site
 router.post('/checkout_action', (req, res, next) => {
     const db = req.app.db;
     const config = req.app.config;
-    const stripeConfig = common.getPaymentConfig();
+    const stripeConfig = getPaymentConfig();
 
     // Create the Stripe payload
     const chargePayload = {
@@ -56,7 +58,7 @@ router.post('/checkout_action', (req, res, next) => {
             orderShipping: req.session.totalCartShipping,
             orderItemCount: req.session.totalCartItems,
             orderProductCount: req.session.totalCartProducts,
-            orderCustomer: common.getId(req.session.customerId),
+            orderCustomer: getId(req.session.customerId),
             orderEmail: req.session.customerEmail,
             orderCompany: req.session.customerCompany,
             orderFirstname: req.session.customerFirstname,
@@ -93,7 +95,7 @@ router.post('/checkout_action', (req, res, next) => {
                     req.session.message = 'Your payment was successfully completed';
                     req.session.paymentEmailAddr = newDoc.ops[0].orderEmail;
                     req.session.paymentApproved = true;
-                    req.session.paymentDetails = '<p><strong>Order ID: </strong>' + newId + '</p><p><strong>Transaction ID: </strong>' + charge.id + '</p>';
+                    req.session.paymentDetails = `<p><strong>Order ID: </strong>${newId}</p><p><strong>Transaction ID: </strong>${charge.id}</p>`;
 
                     // set payment results for email
                     const paymentResults = {
@@ -106,22 +108,22 @@ router.post('/checkout_action', (req, res, next) => {
 
                     // clear the cart
                     if(req.session.cart){
-                        common.emptyCart(req, res, 'function');
+                        emptyCart(req, res, 'function');
                     }
 
                     // send the email with the response
                     // TODO: Should fix this to properly handle result
-                    common.sendEmail(req.session.paymentEmailAddr, 'Your payment with ' + config.cartTitle, common.getEmailTemplate(paymentResults));
+                    sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
                     // redirect to outcome
-                    res.redirect('/payment/' + newId);
+                    res.redirect(`/payment/${newId}`);
                 }else{
                     // redirect to failure
                     req.session.messageType = 'danger';
                     req.session.message = 'Your payment has declined. Please try again';
                     req.session.paymentApproved = false;
-                    req.session.paymentDetails = '<p><strong>Order ID: </strong>' + newId + '</p><p><strong>Transaction ID: </strong>' + charge.id + '</p>';
-                    res.redirect('/payment/' + newId);
+                    req.session.paymentDetails = `<p><strong>Order ID: </strong>${newId}</p><p><strong>Transaction ID: </strong>${charge.id}</p>`;
+                    res.redirect(`/payment/${newId}`);
                 }
             });
         });
@@ -131,7 +133,7 @@ router.post('/checkout_action', (req, res, next) => {
 // Subscription hook from Stripe
 router.all('/subscription_update', async (req, res, next) => {
     const db = req.app.db;
-    const stripeSigSecret = common.getPaymentConfig().stripeWebhookSecret;
+    const stripeSigSecret = getPaymentConfig().stripeWebhookSecret;
     const stripeSig = req.headers['stripe-signature'];
 
     let hook;
@@ -166,7 +168,7 @@ router.all('/subscription_update', async (req, res, next) => {
 
     // Update order status
     await db.orders.updateOne({
-        _id: common.getId(order._id),
+        _id: getId(order._id),
         orderType: 'Subscription'
     }, {
         $set: {
@@ -264,7 +266,7 @@ router.post('/checkout_action_subscription', async (req, res, next) => {
         req.session.message = 'Your subscription was successfully created';
         req.session.paymentEmailAddr = req.body.shipEmail;
         req.session.paymentApproved = true;
-        req.session.paymentDetails = '<p><strong>Order ID: </strong>' + orderId + '</p><p><strong>Subscription ID: </strong>' + subscription.id + '</p>';
+        req.session.paymentDetails = `<p><strong>Order ID: </strong>${orderId}</p><p><strong>Subscription ID: </strong>${subscription.id}</p>`;
 
         // set payment results for email
         const paymentResults = {
@@ -277,14 +279,14 @@ router.post('/checkout_action_subscription', async (req, res, next) => {
 
         // clear the cart
         if(req.session.cart){
-            common.emptyCart(req, res, 'function');
+            emptyCart(req, res, 'function');
         }
 
         // send the email with the response
-        common.sendEmail(req.session.paymentEmailAddr, 'Your payment with ' + config.cartTitle, common.getEmailTemplate(paymentResults));
+        sendEmail(req.session.paymentEmailAddr, `Your payment with ${config.cartTitle}`, getEmailTemplate(paymentResults));
 
         // redirect to outcome
-        res.redirect('/payment/' + orderId);
+        res.redirect(`/payment/${orderId}`);
     });
 });
