+++ lib/testdata.js
@@ -1,4 +1,4 @@
-const { getConfig } = require('./common');
+const { getConfig } = require('./config');
 const { initDb } = require('./db');
 const { fixProductDates, fixDiscountDates, fixProductIds } = require('../lib/testhelper');
 const fs = require('fs');
