+++ public/javascripts/admin.js
@@ -1,4 +1,4 @@
-/* eslint-disable prefer-arrow-callback, no-var, no-tabs */
+/* eslint-disable prefer-arrow-callback, no-var, no-tabs, prefer-template */
 /* globals showNotification, slugify, numeral, moment, feather */
 $(document).ready(function (){
     $.ajaxSetup({
