+++ routes/customer.js
@@ -216,7 +216,7 @@ router.post('/customer/update', async (req, res) => {
             res.status(200).json({ message: 'Customer updated', customer: updatedCustomer.value });
         });
     }catch(ex){
-        console.error(colors.red('Failed updating customer: ' + ex));
+        console.error(colors.red(`Failed updating customer: ${ex}`));
         res.status(400).json({ message: 'Failed to update customer' });
     }
 });
@@ -271,7 +271,7 @@ router.post('/admin/customer/update', restrict, async (req, res) => {
             res.status(200).json({ message: 'Customer updated', customer: updatedCustomer.value });
         });
     }catch(ex){
-        console.error(colors.red('Failed updating customer: ' + ex));
+        console.error(colors.red(`Failed updating customer: ${ex}`));
         res.status(400).json({ message: 'Failed to update customer' });
     }
 });
@@ -296,7 +296,7 @@ router.delete('/admin/customer', restrict, async (req, res) => {
             res.status(200).json({ message: 'Customer deleted' });
         });
     }catch(ex){
-        console.error(colors.red('Failed deleting customer: ' + ex));
+        console.error(colors.red(`Failed deleting customer: ${ex}`));
         res.status(400).json({ message: 'Failed to delete customer' });
     }
 });
@@ -586,7 +586,7 @@ router.post('/customer/reset/:token', async (req, res) => {
         const mailOpts = {
             to: customer.email,
             subject: 'Password successfully reset',
-            body: 'This is a confirmation that the password for your account ' + customer.email + ' has just been changed successfully.\n'
+            body: `This is a confirmation that the password for your account ${customer.email} has just been changed successfully.\n`
         };
 
         // TODO: Should fix this to properly handle result
